Octonions

Documentation for Octonions.

Octonions.extend_analytic

Octonions.extend_analytic — Method

extend_analytic(f, o::Octonion)

Evaluate the extension of the complex analytic function f to the octonions at o.

Given $o = s + a u$, where $s$ is the real part, $u$ is a pure unit octonion, and $a \ge 0$ is the magnitude of the imaginary part of $o$,

\[f(o) = \Re(f(z)) + \Im(f(z)) u,\]

is the extension of f to the octonions, where $z = s + a i$ is a complex analog to $o$.

See [DentoniSce1973] and [ColomboSabadini2020] for details.

DentoniSce1973Dentoni, P. and Sce M. "Funzioni regolari nell'algebra di Cayley." Rendiconti del Seminario matematico della Università di Padova 50 (1973): 251-267. Translation: [ColomboSabadini2020]

ColomboSabadini2020Colombo, F., Sabadini, I., Struppa, D.C. (2020). Regular Functions in the Cayley Algebra. In: Michele Sce's Works in Hypercomplex Analysis. doi: 10.1007/978-3-030-50216-4_6